fixes in comments delete, Stats beta, fixes

+++ src/controllers/user.controller.js
@@ -2,19 +2,23 @@ import {
     registerUser_Service,
     getUserById_Service,
     getUserByEmail_Service,
+    getAdmins_Service,
+    getAllUsers_Service,
     updateUserPassword_Service,
     updateUserInfo_Service,
     deleteUserById_Service,
-    getAdmins_Service
 } from "@/services/user.service";
 
 import { deleteOffertsByUserId_Service } from "@/services/offerts.service";
+import { deleteOffertsCommentByUserId_Controller } from "./comment.controller";
 
 import { saveLogEntrie_Service } from "@/services/log.service";
 
 import { getSession_Controller } from "./auth.controller";
 
 const bcrypt = require('bcrypt');
+
+import { validateEmail } from "@/validations/user.validation";
 
 /**************************{ Create }**************************/
 
@@ -41,7 +45,7 @@ export const registerUser_Controller = async (userData) => {
             actionId: "CREATE",
             actionText: "Se ha registrado en el sistema exitosamente",
         },
-        item:{
+        item: {
             _id: res._id,
             type: "USER",
             name: "",
@@ -58,12 +62,30 @@ export const registerUser_Controller = async (userData) => {
 
 /**************************{ Read }**************************/
 
+export const getTotalUsers_Controller = async () => {
+    try {
+        const res = await getAllUsers_Service();
+
+        const data = res.length;
+    
+        return data;
+    } catch (error) {
+        console.log("getTotalUsers_Controller: ", error);
+        return { error: true, status: 500, message: "Error interno" };
+    };
+};
+
+
 export const getAdmins_Controller = async () => {
-    const res = await getAdmins_Service();
+    try {
+        const res = await getAdmins_Service();
 
-    let data = [];
+        let data = [];
 
-    if (res.length) {
+        if (!res.length) {
+            return data
+        };
+
         res.map((admin) => {
             data.push({
                 name: admin.name,
@@ -75,41 +97,57 @@ export const getAdmins_Controller = async () => {
                 _id: admin._id,
             });
         });
-    };
 
-    return data;
+        return data;
+    } catch (error) {
+        console.log("getAdmins_Controller: ", error);
+        return { error: true, status: 500, message: "Error interno" };
+    };
 };
 
 export const getUserById_Controller = async (id) => {
-    const res = await getUserById_Service(id);
+    try {
+        const res = await getUserById_Service(id);
 
-    const data = {
-        name: res.name,
-        email: res.email,
-        contEmail: res.contEmail,
-        phone: res.phone,
-        role: res.role,
-        active: res.active,
-        _id: res._id,
+        if (!res) return { error: true, status: 404, message: "Ususario no Encontrado" };
+
+        const data = {
+            name: res.name,
+            email: res.email,
+            contEmail: res.contEmail,
+            phone: res.phone,
+            role: res.role,
+            active: res.active,
+            _id: res._id,
+        };
+
+        return data
+    } catch (error) {
+        console.log("getUserById_Controller: ", error);
+        return { error: true, status: 500, message: "Error interno" }
     };
-
-    return data
 };
 
 export const getUserByEmail_Controller = async (email) => {
-    const res = await getUserByEmail_Service(email);
+    try {
+        if (!validateEmail(email)) return { error: true, status: 200, message: "Ingrese un formato de e-mail válido." };
 
-    const data = {
-        name: res.name,
-        email: res.email,
-        contEmail: res.contEmail,
-        phone: res.phone,
-        role: res.role,
-        active: res.active,
-        _id: res._id,
-    };
+        const res = await getUserByEmail_Service(email);
 
-    return data
+        const data = {
+            name: res.name,
+            email: res.email,
+            contEmail: res.contEmail,
+            phone: res.phone,
+            role: res.role,
+            active: res.active,
+            _id: res._id,
+        };
+
+        return data
+    } catch (error) {
+        console.log(error)
+    }
 };
 
 export const getUserWithPasswordByEmail_Controller = async (email) => {
@@ -142,8 +180,8 @@ export const updateUserInfo_Controller = async (userId, data) => {
         const res = await updateUserInfo_Service(userId, data);
 
         const userData = await getSession_Controller();
-        
-        if ( res.name !== userData.user.name) {
+
+        if (res.name !== userData.user.name) {
 
             await saveLogEntrie_Service({
                 action: {
@@ -157,7 +195,7 @@ export const updateUserInfo_Controller = async (userId, data) => {
                 },
                 user: {
                     _id: res._id,
-                    name: userData.user.name 
+                    name: userData.user.name
                 }
             });
 
@@ -194,26 +232,27 @@ export const deleteUserById_Controller = async (userId) => {
         const userData = await getSession_Controller();
 
         const deleteOffertsResponse = await deleteOffertsByUserId_Service(userId);
+        const deletedCommentsResponse = await deleteOffertsCommentByUserId_Controller(userId);
 
-        if (!deleteOffertsResponse.error) {
+        if (!deleteOffertsResponse.error && !deletedCommentsResponse.error) {
             const userDeleted = await getUserById_Controller(userId);
 
             const res = await deleteUserById_Service(userId);
 
-            if ( userId !== userData.user.id) {
+            if (userId !== userData.user.id) {
                 await saveLogEntrie_Service({
                     action: {
                         actionId: "DELETE",
                         actionText: "eliminó al usuario de nombre",
                     },
-                    item:  {
+                    item: {
                         _id: userDeleted._id,
                         type: "USER",
                         name: `"${userDeleted.name}", junto con todas sus ofertas`,
                     },
                     user: {
                         _id: userData.user.id,
-                        name: userData.user.name 
+                        name: userData.user.name
                     }
                 });
             } else {
@@ -233,7 +272,7 @@ export const deleteUserById_Controller = async (userId) => {
                     }
                 });
             };
-            
+
             return res;
         };
     } catch (error) {
